Refactor main and move compile logic
